feat: Apply new MySQL schema and remove local authentication
This commit applies the new MySQL schema to the entire codebase,
including domain objects (VOs), DTOs, and MyBatis mapper XML files.

Key changes include:
- Updated table and column names to match the new schema (e.g., , , , , , , , ).
- Adjusted field names and types in Java VO and DTO classes accordingly.
- Modified MyBatis mapper XML files to reflect the new table and column structures, and updated SQL queries (INSERT, UPDATE, DELETE, SELECT).
- Removed local authentication components, including , , , , and , as the application now relies solely on external authentication (Kakao).
- Updated  to use  database and  user.
- Refactored  to be a Spring-aware test, using  bean for connection.
- Removed outdated security-related test files (, ).

These changes ensure the backend aligns with the new database structure and simplifies the authentication mechanism.

+++ src/main/java/org/scoula/asset/domain/AssetInfoVO.java
@@ -5,15 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class AssetInfoVO {
     private String email;
-    private BigDecimal asset;
+    private Long asset;
     private String segment;
     private String filename1;
     private String filename2;
